refactor: extract DeleteButton into a separate component
