+++ src/features/transactionTable/components/columns/columns.tsx
@@ -2,20 +2,9 @@ import { ColumnDef } from '@tanstack/react-table';
 import { TableColumnHeader } from '@/components/TableColumnHeader';
 import { TransactionTableData } from './types';
 import { Button } from '@/components/ui/button';
-import { Edit as EditIcon, Trash as TrashIcon } from 'lucide-react';
+import { Edit as EditIcon } from 'lucide-react';
 import { AddTransactionModal } from '@/components/AddTransactionModal';
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from '@/components/ui/alert-dialog';
-import { toast } from '@/hooks/use-toast';
+import { DeleteButton } from './DeleteButton';
 
 export const columns: ColumnDef<TransactionTableData>[] = [
   {
@@ -161,36 +150,6 @@ export const columns: ColumnDef<TransactionTableData>[] = [
     cell: ({ row }) => {
       const transactionToEdit = row.original;
 
-      const handleDeleteTransaction = () => {
-        try {
-          const existingTransactions = JSON.parse(
-            localStorage.getItem('transactions') || '[]',
-          ) as TransactionTableData[];
-
-          const updatedTransactions = existingTransactions.filter(
-            (transaction) => transaction.id !== transactionToEdit.id,
-          );
-
-          localStorage.setItem(
-            'transactions',
-            JSON.stringify(updatedTransactions),
-          );
-
-          toast({
-            title: 'Transakce byla úspěšně smazána',
-            duration: 5000,
-            className: 'bg-green-100 border-green-500 text-green-900',
-          });
-        } catch (error) {
-          console.error('Error deleting transaction:', error);
-
-          toast({
-            variant: 'destructive',
-            title: 'Nastala chyba při mazání transakce',
-          });
-        }
-      };
-
       return (
         <div className="flex justify-center space-x-2">
           <AddTransactionModal transactionToEdit={transactionToEdit}>
@@ -199,29 +158,7 @@ export const columns: ColumnDef<TransactionTableData>[] = [
             </Button>
           </AddTransactionModal>
 
-          <AlertDialog>
-            <AlertDialogTrigger asChild>
-              <Button variant="ghost" size="icon">
-                <TrashIcon className="h-4 w-4 text-red-500" />
-              </Button>
-            </AlertDialogTrigger>
-            <AlertDialogContent>
-              <AlertDialogHeader>
-                <AlertDialogTitle>
-                  Opravdu chcete smazat tuto transakci?
-                </AlertDialogTitle>
-                <AlertDialogDescription>
-                  Transakce bude trvale odstraněna.
-                </AlertDialogDescription>
-              </AlertDialogHeader>
-              <AlertDialogFooter>
-                <AlertDialogCancel>Zrušit</AlertDialogCancel>
-                <AlertDialogAction onClick={handleDeleteTransaction}>
-                  Smazat
-                </AlertDialogAction>
-              </AlertDialogFooter>
-            </AlertDialogContent>
-          </AlertDialog>
+          <DeleteButton transactionToEdit={transactionToEdit} />
         </div>
       );
     },
